requirements.txt
# Parquet Visualizer - Python Dependencies

# Core Parquet handling
pyarrow>=14.0.0
pandas>=2.0.0

# Web framework
streamlit>=1.29.0

# Image handling
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0  # Property-based testing

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
